+++ CHANGELOG.md
@@ -15,6 +15,8 @@ Polyverse Boost ReST API (Backend)
 - Refreshing Data References ( /api/user_project/{org}/{project}/data_references (POST) ) will check if known OpenAI files are missing and re-upload them
 - Ensure Project Status returns Out of Sync if AI files are missing (e.g. uploaded to OpenAI in past, but OpenAI reports they are missing)
 - Ensure changing project data - via PATCH, POST or PUT resets the generators to recreate the resources
+- Deleting a project will try to delete all sub-resources as well (e.g. status, data_references, resources, etc)
+- Deleting a Project resource will try to delete all generator data as well
 
 ### Bug Fixes
 - Fix OpenAI throttling issue with Assistant batch delete (wasn't waiting 1 second)